requirements.txt
fastapi
blarify
gitpython
uvicorn